对微信支付V3的MarkingApis使用partial类
