+++ src/Senparc.Weixin.TenPay/Senparc.Weixin.TenPayV3.Test/Apis/Marketing/MarketingPartnershipsApisTests.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Senparc.CO2NET.Extensions;
+using Senparc.Weixin.Helpers;
+using Senparc.Weixin.TenPayV3.Apis.Marketing;
+using Senparc.Weixin.TenPayV3.Entities;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Senparc.Weixin.TenPayV3.Apis.Tests
+{
+    public partial class MarketingApisTests
+    {
+        #region 委托营销接口
+
+
+        #endregion
+
+    }
+}
